fixes serverside components
