+++ frontend/app/dashboard/layout.tsx
@@ -1,5 +1,3 @@
-'use client'
-
 import { NavBar } from 'components/navigations/NavBar'
 
 export default function DashboardLayout({ children }: { children: React.ReactNode }) {
